v1.8 - August 24th, 2014
v1.8 - August 24th, 2014
o Fixed some undefined errors found in ManageSplitForums.php.
o Fixed issue with Manage Boards UI installed where copying blocks was selected by default.
o Modified code so that domain names with and without www. are Properly recognized.
o Rewrote function that copies SimplePortal blocks from one subforum to another.
o Fixed bug with SimplePortal saving configuration settings because variable wasn't globalized.
o Fixed variable type-casting bug with deleting SimplePortal blocks from a subforum.
o Fixed mod installer issue for people without SimplePortal installed.

+++ ManageSplitForums.php
@@ -154,8 +154,7 @@ Function EditSubForum($sub)
 			$modSettings['subforum_modify_sp_portal'] = $modSettings['sp_portal_mode'];
 		if (!isset($modSettings['subforum_modify_sp_standalone']))
 			$modSettings['subforum_modify_sp_standalone'] = $modSettings['sp_standalone_url'];
-		if ($_REQUEST['sa'] == 'newsub')
-			$modSettings['subforum_modify_sp_blocks'] = -2;
+		$modSettings['subforum_modify_sp_blocks'] = ($_REQUEST['sa'] == 'newsub' ? -2 : -1);
 		$txt['subforum_modify_sp_portal'] = $txt['sp_portal_mode'];
 		$txt['subforum_modify_sp_standalone'] = $txt['sp_standalone_url'];
 	}
@@ -226,12 +225,12 @@ function SaveSubForum($sub)
 	add_subforum($arr);
 
 	// Populate the Simple Portal blocks for this subforum:
-	if ($_POST['subforum_modify_sp_blocks'] == -2)
+	if (isset($_POST['subforum_modify_sp_blocks']) && $_POST['subforum_modify_sp_blocks'] == -2)
 	{
 		delete_sp_blocks($arr['forumid']);
 		use_default_sp_blocks($arr['forumid']);
 	}
-	elseif ($_POST['subforum_modify_sp_blocks'] > -1)
+	elseif (isset($_POST['subforum_modify_sp_blocks']) && $_POST['subforum_modify_sp_blocks'] > -1)
 	{
 		delete_sp_blocks($arr['forumid']);
 		copy_sp_blocks($_POST['subforum_modify_sp_blocks'], $arr['forumid']);
